Create psycud
I want to create a budget tracking application. As I am new to coding, could you help me out by writing python code for the same. The requirements that I have are as follows: 1) At the entry screen, give the user an option to enter their name and email address. Store this data properly in a database. Create a UI for this with a button which leads them to the next step. 2) Post this, allow the user to enter their monthly budget and savings goal. After the user has entered this information, this information gets locked for the current month. Store this information in the database as well. However, this can be edited for every new month of the year. Create a UI for this with a button which leads them to the next step. 3) Now, allow the user to input their expense and allow them to assign a category to the same. Save this information in the database. 4) The user should have the capability to add multiple expense items after clicking the button add expense and all of these should be saved in the database. 5) Show all this information in a tabular format and automatically update the table whenever a new expense is created. 6) There should be different tabs for different months.